pm/x/pm/keeper/msg_server_aggregator_test.go
package keeper_test

import (
	"testing"

	sdk "github.com/cosmos/cosmos-sdk/types"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"

	keepertest "github.com/supsi-dacd-isaac/parity-cosmos/pm/testutil/keeper"
	"github.com/supsi-dacd-isaac/parity-cosmos/pm/x/pm/keeper"
	"github.com/supsi-dacd-isaac/parity-cosmos/pm/x/pm/types"
)

func TestAggregatorMsgServerCreate(t *testing.T) {
	k, ctx := keepertest.PmKeeper(t)
	srv := keeper.NewMsgServerImpl(*k)
	wctx := sdk.WrapSDKContext(ctx)
	creator := "A"
	expected := &types.MsgCreateAggregator{Creator: creator}
	_, err := srv.CreateAggregator(wctx, expected)
	require.NoError(t, err)
	rst, found := k.GetAggregator(ctx)
	require.True(t, found)
	require.Equal(t, expected.Creator, rst.Creator)
}

func TestAggregatorMsgServerUpdate(t *testing.T) {
	creator := "A"

	for _, tc := range []struct {
		desc    string
		request *types.MsgUpdateAggregator
		err     error
	}{
		{
			desc:    "Completed",
			request: &types.MsgUpdateAggregator{Creator: creator},
		},
		{
			desc:    "Unauthorized",
			request: &types.MsgUpdateAggregator{Creator: "B"},
			err:     sdkerrors.ErrUnauthorized,
		},
	} {
		t.Run(tc.desc, func(t *testing.T) {
			k, ctx := keepertest.PmKeeper(t)
			srv := keeper.NewMsgServerImpl(*k)
			wctx := sdk.WrapSDKContext(ctx)
			expected := &types.MsgCreateAggregator{Creator: creator}
			_, err := srv.CreateAggregator(wctx, expected)
			require.NoError(t, err)

			_, err = srv.UpdateAggregator(wctx, tc.request)
			if tc.err != nil {
				require.ErrorIs(t, err, tc.err)
			} else {
				require.NoError(t, err)
				rst, found := k.GetAggregator(ctx)
				require.True(t, found)
				require.Equal(t, expected.Creator, rst.Creator)
			}
		})
	}
}

func TestAggregatorMsgServerDelete(t *testing.T) {
	creator := "A"

	for _, tc := range []struct {
		desc    string
		request *types.MsgDeleteAggregator
		err     error
	}{
		{
			desc:    "Completed",
			request: &types.MsgDeleteAggregator{Creator: creator},
		},
		{
			desc:    "Unauthorized",
			request: &types.MsgDeleteAggregator{Creator: "B"},
			err:     sdkerrors.ErrUnauthorized,
		},
	} {
		t.Run(tc.desc, func(t *testing.T) {
			k, ctx := keepertest.PmKeeper(t)
			srv := keeper.NewMsgServerImpl(*k)
			wctx := sdk.WrapSDKContext(ctx)

			_, err := srv.CreateAggregator(wctx, &types.MsgCreateAggregator{Creator: creator})
			require.NoError(t, err)
			_, err = srv.DeleteAggregator(wctx, tc.request)
			if tc.err != nil {
				require.ErrorIs(t, err, tc.err)
			} else {
				require.NoError(t, err)
				_, found := k.GetAggregator(ctx)
				require.False(t, found)
			}
		})
	}
}
